requirements.txt
fastapi==0.115.4
uvicorn==0.35.0
opencv-python>=4.10.0,<4.11.0
pillow==10.4.0
numpy>=1.23.5,<2.0.0
scikit-image==0.24.0
exifread==3.0.0

torch>=2.1.0
torchvision
pillow
numpy
scikit-learn
matplotlib
opencv-python
jsonlines

